+++ client/src/lib/constants.js
@@ -1,4 +1,8 @@
 const { RegexButWithWords } = require("regex-but-with-words");
+const mixedTemplate = require("./docs/templates/slz-mixed.json");
+const vsiTemplate = require("./docs/templates/slz-vsi.json");
+const vsiEdgeTemplate = require("./docs/templates/slz-vsi-edge.json");
+const powerTemplate = require("./docs/templates/power-sap-hana.json");
 
 module.exports = {
   varDotRegion: "${var.region}",
@@ -793,5 +797,86 @@ module.exports = {
       "Tier1-Flash-3",
     ],
     tor01: ["Tier3-Flash-2", "Tier3-Flash-1", "Tier1-Flash-2", "Tier1-Flash-1"],
+  },
+  templates: ["Mixed", "VSI", "VSI Edge", "Power VS SAP Hana"],
+  template_dropdown_map: {
+    Mixed: {
+      template: mixedTemplate,
+      name: "Landing Zone Mixed Pattern",
+      image: "",
+      patternDocText:
+        "A default template based on the IBM Landing Zone Mixed Pattern. This is the default pattern for CRAIG.",
+      includes: [
+        "A resource group for cloud services and for each VPC",
+        "A management and workload VPC connected by a transit gateway",
+        "A flow log collector for each VPC",
+        "Object storage instances for flow logs and activity tracker",
+        "Encryption keys in either a Key Protect or Hyper Protect Crypto Services instance",
+        "All necessary networking rules to allow communication",
+        "Virtual Private endpoints for Cloud Object storage in each VPC",
+        "A Red Hat OpenShift cluster in the workload VPC",
+        "An example Virtual Server instance deployment in the management VPC",
+        "A VPN Gateway in the Management VPC",
+      ],
+    },
+    VSI: {
+      template: vsiTemplate,
+      name: "Landing Zone VSI Pattern",
+      image: "",
+      patternDocText:
+        "Based on the IBM Landing Zone VSI Pattern, deploys an example application server deployment in both the Management and Workload VPC.",
+      includes: [
+        "A resource group for cloud services and for each VPC",
+        "A management and workload VPC connected by a transit gateway",
+        "A flow log collector for each VPC",
+        "Object storage instances for flow logs and activity tracker",
+        "Encryption keys in either a Key Protect or Hyper Protect Crypto Services instance",
+        "All necessary networking rules to allow communication",
+        "Virtual Private endpoints for Cloud Object storage in each VPC",
+        "An example Virtual Server deployment in the management VPC",
+        "An example Virtual Server deployment in the workload VPC",
+        "A VPN Gateway in the Management VPC",
+      ],
+    },
+    "VSI Edge": {
+      template: vsiEdgeTemplate,
+      name: "Landing Zone VSI Edge Pattern",
+      image: "",
+      patternDocText:
+        "Based on the IBM Landing Zone VSI Edge Pattern, deploys an Edge VPC with one VSI and an F5 Big IP instance with VPN and WAF.",
+      includes: [
+        "A resource group for cloud services and for each VPC",
+        "A management and workload VPC connected by a transit gateway",
+        "An Edge VPC with F5 Big IP and needed network interfaces",
+        "A flow log collector for each VPC",
+        "Object storage instances for flow logs and activity tracker",
+        "Encryption keys in either a Key Protect or Hyper Protect Crypto Services instance",
+        "All necessary networking rules to allow communication",
+        "Virtual Private endpoints for Cloud Object storage in each VPC",
+        "An example Virtual Server deployment in the management VPC",
+        "An example Virtual Server deployment in the workload VPC",
+        "A VPN Gateway in the Management VPC",
+      ],
+    },
+    "Power VS SAP Hana": {
+      template: powerTemplate,
+      name: "Power VS SAP Hana Pattern",
+      image: "",
+      patternDocText:
+        "Based on existing SAP Hana solutions, this template creates base infrastructure with SAP system landscape that leverages the services from the VPC landing zone as well as the needed components to get started with a Power VS Virtual Server environment",
+      includes: [
+        "A resource group for cloud services and for each VPC",
+        "A management and workload VPC connected by a transit gateway",
+        "An Edge VPC with F5 Big IP and needed network interfaces",
+        "A Power VS Instances for SAP Hana, SAP Netweaver, and Secure File Share.",
+        "A flow log collector for each VPC",
+        "Object storage instances for flow logs and activity tracker",
+        "Encryption keys in either a Key Protect or Hyper Protect Crypto Services instance",
+        "All necessary networking rules to allow communication",
+        "Virtual Private endpoints for Cloud Object storage in each VPC",
+        "An example Virtual Server instance deployment in the management VPC",
+        "A VPN Gateway in the Management VPC",
+      ],
+    },
   },
 };